review fixes, gradle added, editorconfig added, reformatting and refactoring
