+++ README.md
@@ -1,1 +1,3 @@
 # Working-with-Files-in-Java-IO-NIO
+
+Specify main class in the `build.gradle.kts` and run with `gradle run`
