+++ app/src/main/java/task3/FastFileMover.java
@@ -5,6 +5,6 @@ import java.io.IOException;
 @FunctionalInterface
 public interface FastFileMover {
 
-    void moveFile(String sourcePath, String destinationPath) throws IOException;
+  void moveFile(String sourcePath, String destinationPath) throws IOException;
 
 }
